PhishTank
This software created with C# and Winforms performs a web request on the
phish tank website using an Api-Key. It downloads the database
containing suspected or verified phishing sites in XML format and parses
it into a DataGrid. The software allows you to search the XML Database
Via LINQ query to Xml Database